azure-cnci-functions/requirements.txt
azure-functions
azure-devops
slackclient
requests